change user errors

+++ app/controller/user_contoller.py
@@ -11,8 +11,8 @@ def add_user():
     if request.method == "POST":
         data = request.get_json()
         password_from_data = data.get("password")
-        if not password_from_data:
-            return jsonify({"message": "Password is required"}), 400
+        if type(password_from_data) != str:
+            return jsonify({"message": "Password must be string"}), 400
         if len(password_from_data) < 6:
             return jsonify({"message": "Password must be at least 6 characters long"}), 400
         try:
@@ -41,7 +41,7 @@ def add_user():
         except Exception as e:
             db.session.rollback()
             return jsonify({"message":"An unexpected error occurred"}) , 500
-        return jsonify({"message": "User added successfully"}), 201
+        return jsonify({"message": "User added successfully"}), 200
 
 @user_bp.route('/<int:user_id>', methods=['DELETE'])
 def delete_user(user_id):
@@ -52,11 +52,25 @@ def delete_user(user_id):
         data = request.get_json()
         email_val = data.get("email")
         password_val = data.get("password")
+    try:
+        if type(password_val) != str:
+            return jsonify({"message":"Password must be srting"}) , 400
         if user and user.check_password(password_val) and user.email == email_val:
             db.session.delete(user)
             db.session.commit()
             return jsonify({"message": "User deleted successfully"}), 200
         return jsonify({"message": "Invalid email or password"}), 401
+    except IntegrityError as e:
+        db.session.rollback()
+        error_msg = str(e.orig)
+        if "ck_user_email_format" in error_msg:
+            return jsonify({"message":"Invalid email"}), 400
+    except ValueError as e:
+        db.session.rollback()
+        return jsonify({"message": str(e)}), 400
+    except Exception as e:
+            db.session.rollback()
+            return jsonify({"message":"An unexpected error occurred"}) , 500
 
 @user_bp.route('/<int:user_id>', methods=['PATCH'])
 def update_user(user_id):
@@ -64,15 +78,38 @@ def update_user(user_id):
     if not user:
         return jsonify({"message": "User not found"}), 404
     data = request.get_json()
-    if "username" in data:
-        user.username = data.get("username")
-    if "email" in data:
-        user.email = data.get("email")
-    if "age" in data:
-        user.age = data.get("age")
-    if "password" in data:
-        user.set_password(data.get("password")) 
-    db.session.commit()
+    try:
+        if "username" in data:
+            user.username = data.get("username")
+        if "email" in data:
+            user.email = data.get("email")
+        if "age" in data:
+            user.age = data.get("age")
+        if "password" in data:
+            if type(data.get("password")) != str:
+                return jsonify({"message":"Password must be srting"}) , 400
+            if len(data.get("password")) < 6:
+                return jsonify({"message":"Password must be at least 6 characters long"}), 400
+            else:
+                user.set_password(data.get("password"))  
+        db.session.commit()
+    except IntegrityError as e:
+        db.session.rollback()
+        error_msg = str(e.orig)
+        if "uq_user_email" in error_msg:
+            return jsonify({"message": "This email is already registered"}) , 400
+        if "ck_user_age" in error_msg:
+            return jsonify({"message":"Age must be between 5 and 100"}) , 400
+        if "ck_user_username_not_empty" in error_msg:
+            return jsonify({"message":"Username cannot be empty"}) , 400
+        if "ck_user_email_format" in error_msg:
+            return jsonify({"message":"Invalid email"}), 400
+    except ValueError as e:
+        db.session.rollback()
+        return jsonify({"message": str(e)}), 400
+    except Exception as e:
+            db.session.rollback()
+            return jsonify({"message":"An unexpected error occurred"}) , 500
     return jsonify({"message": "User updated successfully"}), 200
 
 @user_bp.route('/<int:user_id>' , methods=['GET'])
@@ -95,16 +132,36 @@ def change_user(user_id):
     user = User.query.get(user_id)
     if not user:
         return jsonify({"message": "User not found"}), 404
-    required_fields = ["username", "email", "age", "password"]
-    missing_fields = [field for field in required_fields if field not in data]
-    if missing_fields:
-        return jsonify({"error": "Missing data",
-            "message": f"Fields required: {', '.join(missing_fields)}"}), 400
-    if len(data["password"]) < 6:
-        return jsonify({"message": "Password must be at least 6 characters long"}), 400
-    else:user.set_password(data.get("password"))
-    user.username = data.get("username")
-    user.email = data.get("email")
-    user.age = data.get("age")
-    db.session.commit()
+    try:
+        required_fields = ["username", "email", "age", "password"]
+        missing_fields = [field for field in required_fields if field not in data]
+        if missing_fields:
+            return jsonify({"error": "Missing data",
+                "message": f"Fields required: {', '.join(missing_fields)}"}), 400
+        if type(data.get("password")) != str:
+                return jsonify({"message":"Password must be srting"}) , 400
+        if len(data.get("password")) < 6:
+            return jsonify({"message": "Password must be at least 6 characters long"}), 400
+        user.set_password(data.get("password"))
+        user.username = data.get("username")
+        user.email = data.get("email")
+        user.age = data.get("age")
+        db.session.commit()
+    except IntegrityError as e:
+        db.session.rollback()
+        error_msg = str(e.orig)
+        if "uq_user_email" in error_msg:
+            return jsonify({"message": "This email is already registered"}) , 400
+        if "ck_user_age" in error_msg:
+            return jsonify({"message":"Age must be between 5 and 100"}) , 400
+        if "ck_user_username_not_empty" in error_msg:
+            return jsonify({"message":"Username cannot be empty"}) , 400
+        if "ck_user_email_format" in error_msg:
+            return jsonify({"message":"Invalid email"}), 400
+    except ValueError as e:
+        db.session.rollback()
+        return jsonify({"message": str(e)}), 400
+    except Exception as e:
+            db.session.rollback()
+            return jsonify({"message":"An unexpected error occurred"}) , 500
     return jsonify({"message": "User information changed successfully"}), 200
